requirements.txt
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
shiboken6==6.6.1
PyYAML~=6.0.1